fix(migration): rate-limit columns BIGINT to match GORM AutoMigrate; document limiter RPM/TPM ordering
Full-flow adversarial review (5 Sonnet-5 reviewers + per-finding verify against
a clean worktree, each finding re-checked against source) surfaced two real,
independently-confirmed schema-convergence defects and one design nuance.

- 026 + 023: rpm_limit/tpm_limit declared SQL INT, but GORM maps the plain Go
  `int` fields (entity.ModelRateLimit / entity.Tenant / repo Token) to postgres
  BIGINT — verified empirically against real PG (AutoMigrate yields bigint). The
  SQL/AutoMigrate divergence is masked on normal boots (AutoMigrate runs first)
  but on a runner-first DB (DR restore) the next boot would rewrite the column
  integer→bigint under an ACCESS EXCLUSIVE lock — heavy on the large tokens
  table. Declaring BIGINT makes both creation paths identical, honouring each
  migration's own convergence contract. No-op on already-applied DBs.
- business_model_rate_limit.go: documented that RPM admission is recorded before
  the TPM check (uniform with the token/tenant dimensions) so sustained TPM
  overage also consumes RPM budget — a deliberate, in-scope-documented design
  choice, not a defect; and the (tenant,model) key's safety invariant (tenant
  IDs are system-generated, never contain ':').

Verified: migration PG suite green against disposable postgres:15; middleware
builds. Column-type divergence confirmed + fix validated with a throwaway real-PG
AutoMigrate probe (deleted).

+++ migrations/026_create_model_rate_limits.sql
@@ -1,6 +1,6 @@
 -- 026_create_model_rate_limits.sql
 -- Idempotent PG-only creation of model_rate_limits: one row per
--- (tenant_id, model) carrying rpm_limit / tpm_limit (INT NOT NULL DEFAULT 0,
+-- (tenant_id, model) carrying rpm_limit / tpm_limit (BIGINT NOT NULL DEFAULT 0,
 -- 0 = unlimited) — the model dimension of the business rate limiter.
 --
 -- WHY: migration 023 gave the token and tenant dimensions their limit
@@ -26,13 +26,18 @@
 --     the AutoMigrate-first ordering are both no-ops.
 --   * Index names match the GORM uniqueIndex tag (uk_model_rate_limits_
 --     tenant_model) so the two creation paths converge on identical schema.
+--   * rpm_limit/tpm_limit are BIGINT, not INT: GORM maps a plain Go `int`
+--     field (entity.ModelRateLimit) to postgres bigint, so declaring INT here
+--     would DIVERGE from AutoMigrate — on a runner-first DB (DR restore) the
+--     next boot's AutoMigrate would then rewrite the column (integer→bigint)
+--     under an ACCESS EXCLUSIVE lock. BIGINT makes both paths identical.
 
 CREATE TABLE IF NOT EXISTS model_rate_limits (
     id         BIGSERIAL PRIMARY KEY,
     tenant_id  VARCHAR(36)  NOT NULL DEFAULT 'default',
     model      VARCHAR(255) NOT NULL,
-    rpm_limit  INT          NOT NULL DEFAULT 0,
-    tpm_limit  INT          NOT NULL DEFAULT 0,
+    rpm_limit  BIGINT       NOT NULL DEFAULT 0,
+    tpm_limit  BIGINT       NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ,
     updated_at TIMESTAMPTZ
 );
